Rewrite headerRewriter() to not insert spurious line breaks
headerRewriter() was called on a buffer filled by a tee reader
as part of parsing the message headers. This buffer does not
necessarily end with a newline (or in the headers; it might
contain part of the body), but headerRewrite() was inserting
one after each line it read (including a final one not terminated
by it).

Depending on the email, this can cause DKIM signature verification
failures as it inserts an additional newline into the message body,
thus breaking the hash.

This rewrites the logic to just handle the first line "From" stripping
and insertion of extra headers directly. There's some nasty business
involved in stripping the first newline to account for various line
endings.

This might still end up with mixed line endings, but that has been
the case before anyway, since we only parsed the headers. It seems
postfix always passes us '\n', though.

Fixes #5